- Обновил UI-тесты; - Доделал рефакторинг кода ProfileViewController; - Перенес показ алерта о выходе из профиля в ProfileViewController.
